+++ sdk/search/search-documents/recordings/node/searchindexerclient_indexers/recording_throws_error_for_invalid_indexer_object.js
@@ -1,6 +1,6 @@
 let nock = require('nock');
 
-module.exports.hash = "9ed1339e7be399a8d5107634941d8299";
+module.exports.hash = "b546af201e9f4b3ed9f7fb66f796c57e";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
@@ -18,7 +18,7 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Expires',
   '-1',
   'request-id',
-  'c61f526e-0535-4733-a646-33bd22f86653',
+  '907e040a-e89e-4064-b2ed-f80b5559d003',
   'elapsed-time',
   '6',
   'OData-Version',
@@ -28,6 +28,6 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Strict-Transport-Security',
   'max-age=15724800; includeSubDomains',
   'Date',
-  'Wed, 26 Aug 2020 01:11:08 GMT',
+  'Sat, 26 Sep 2020 18:12:34 GMT',
   'Content-Length',
   '95' ]);